Added myself as a developer in the platform's pom.xml file.
Using the Hadoop subproject as a guide, created a skeletal subproject for the MySQL integration necessary to make the appliance work.  This involved a very small number of edits to the platform's pom.xml file, adding only two lines, and a lot of copying and pasting from the Hadoop subproject configuration.  The project builds just fine, but since I do not know the run procedure yet for skye, I am unable to verify that it runs correctly.

Created the skeleton for the MySqlObjectMetadataRepository class, in order to test the subproject creation.

If I've broken anything, it is important to note that corporal punishment is illegal in the state of North Carolina:)

~ Joshua Soles <josh@infobelt.com>