Completed the conversion methods between JPAArchiveContentBlock and ArchiveContentBlock types, complete with Unit Testing.
Wrote the methods to convert between JPAObjectMetadata and ObjectMetadata (tests will be written soon).

Wrote methods on the JPAObjectMetadataRepository which write persistent JPAObjectMetadata to the datastore, as well as some of the retrieval methods.

Removed the Set fields from the ObjectSet, as this list is intended to be store primarily on disk and accessed record by record by the OMR.

Fixed an issue in which the StoreRegistry necessary to translate between JPAArchiveContentBlocks and ArchiveContentBlocks was not injected properly.

~ Joshua Soles <josh@infobelt.com>