Added a name parameter to all methods which create ObjectSets, so that names can be set prior to any file system writes that may occur.
Added the JPAObjectSet class, which is a persistable version of the ObjectSet.

Since the requirements for ArchiveContentBlocks have changed, it is no longer necessary to inject the StoreRegistry into a JPAArchiveContentBlock.  In a later commit, comments to the contrary will be removed.

Since these requirements were changed, it is no longer necessary to test the ArchiveContentBlock conversions.

Temporarily removed a key signing requirement which caused the Skye API not to build properly, pending a solution.

~ Joshua Soles <josh@infobelt.com>