Added an interface which all domain objects will implement called Identifiable.  It indicates that an object can use getId().
Created new RuntimeException classes which indicate that either a persistable entity is found, and it should not have been, or that it was expected and was not found.

Given that there were two AuditLogDAO classes, and one of them was incorrectly named and written, the offending duplicate class has been removed.

Added the InformationStoreDefinitionDAO and the InformationStoreDefinitionDomainResource classes, which perform their respective tasks for the InformationStoreDefinition.

Changed the JPAObjectMetadata field relating to the informationStoreId.

Finished the (untested) functionality on the JPAObjectMetadataRepository.

Added the JPAObjectSet as a persistable entity in the persistence.xml file.

Added the create and update methods to the DAO objects, which check for the appropriate conditions before creating or updating an Entity.

~ Joshua Soles <josh@infobelt.com>