Added the ArchiveStoreInstanceDAO and TaskStatisticsDAO, along with the associated REST API Resources.
Fixed the JPAObjectMetadataRepositoryTest, and found that one must commit to the datastore prior to performing queries more complicated than the find function on them.

Modified the JPAObjectMetadata class and metamodel to include new fields from a recent commit.

Fixed some bugs in the AbstractPaginatingDAO class.

Added H2 as a dependency to the core package, for JPA testing.

~ Joshua Soles <josh@infobelt.com>