Implemented pagination support in search API and OMS objects.
Implemented methods which count total results for a query prior to executing it.

Implemented CLI support for searches with pagination or which returns all results, both with and without a project as part of the search.

Added (commented out) lines to the skye-sqlserver.yml and skye-sqlserver-worker.yml files which would allow searches to be run against the test server's search client (atpblade05.infobelt.com).  Simply uncomment them to use this server in local searches.