Added the /objects/search/count API endpoints, which count search results.  Note that right now, counting without specifying a project id will result in a count that does not consider project search permissions.  This issue is being investigated, and may result in inaccurate pagination information.
Added the objects --searchCount CLI command, which counts the results a query would produce before it is completed, with an optional project constraint on the search.

Added a line to the objects --search output which displays the page number of the result.  If no pageNumber is specified, this result may be 0 in error.  Fix forthcoming.